refactor: 도메인별 엔티티 폴더 생성
