--- src/main/java/re/kr/icuh/icuhplatformadmin/domain/FileStatus.java
@@ -1,10 +1,0 @@
-package re.kr.icuh.icuhplatformadmin.domain;
-
-public enum FileStatus {
-    PENDING,
-    APPROVED,
-    REJECTED,
-    DELETED,
-    UPDATED_PENDING,
-    DELETED_PENDING
-}
