+++ src/main/java/re/kr/icuh/icuhplatformadmin/domain/article/ArticleEditRequest.java
@@ -1,9 +1,10 @@
-package re.kr.icuh.icuhplatformadmin.domain;
+package re.kr.icuh.icuhplatformadmin.domain.article;
 
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.testcontainers.shaded.com.google.common.hash.Hashing;
+import re.kr.icuh.icuhplatformadmin.domain.file.FileEditRequest;
 
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
