+++ src/main/java/re/kr/icuh/icuhplatformadmin/domain/article/ArticleStatus.java
@@ -1,0 +1,11 @@
+package re.kr.icuh.icuhplatformadmin.domain.article;
+
+public enum ArticleStatus {
+    PENDING,
+    APPROVED,
+    REJECTED,
+    DELETED,
+    UPDATED_PENDING,
+    UPDATED_APPROVED,
+    DELETED_PENDING
+}
