+++ src/main/java/re/kr/icuh/icuhplatformadmin/dto/ArticleApprovedResponse.java
@@ -1,8 +1,8 @@
 package re.kr.icuh.icuhplatformadmin.dto;
 
-import re.kr.icuh.icuhplatformadmin.domain.Article;
-import re.kr.icuh.icuhplatformadmin.domain.ArticleStatus;
-import re.kr.icuh.icuhplatformadmin.domain.FileStatus;
+import re.kr.icuh.icuhplatformadmin.domain.article.Article;
+import re.kr.icuh.icuhplatformadmin.domain.article.ArticleStatus;
+import re.kr.icuh.icuhplatformadmin.domain.file.FileStatus;
 
 import java.time.LocalDateTime;
 import java.util.List;
