+++ src/main/java/re/kr/icuh/icuhplatformadmin/service/ArticleService.java
@@ -3,10 +3,10 @@ package re.kr.icuh.icuhplatformadmin.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import re.kr.icuh.icuhplatformadmin.domain.Article;
-import re.kr.icuh.icuhplatformadmin.domain.ArticleEditRequest;
-import re.kr.icuh.icuhplatformadmin.domain.ArticleStatus;
-import re.kr.icuh.icuhplatformadmin.domain.FileStatus;
+import re.kr.icuh.icuhplatformadmin.domain.article.Article;
+import re.kr.icuh.icuhplatformadmin.domain.article.ArticleEditRequest;
+import re.kr.icuh.icuhplatformadmin.domain.article.ArticleStatus;
+import re.kr.icuh.icuhplatformadmin.domain.file.FileStatus;
 import re.kr.icuh.icuhplatformadmin.dto.ArticleListResponse;
 import re.kr.icuh.icuhplatformadmin.dto.ArticlePendingResponse;
 import re.kr.icuh.icuhplatformadmin.repository.ArticleQueryRepository;
